make login and jwtSign middleware
