+++ routes/auth/POST-login.js
@@ -1,18 +1,8 @@
-const jwt = require('jsonwebtoken');
-
-const login = require('../../util/login');
+const login = require('../../middleware/login');
+const jwtSign = require('../../middleware/jwtSign');
 
 module.exports = (app) => {
-  app.post('/auth/login', (req, res) => {
-    login(req, res, (email) => {
-      const jwtExpiry = 24 * 60 * 60; // 1 day in seconds
-      const token = jwt.sign({ email }, process.env.JWT_KEY, {
-        algorithm: 'HS256',
-        expiresIn: jwtExpiry,
-      });
-
-      res.cookie('token', token, { maxAge: jwtExpiry * 1000 });
-      return res.status(200).json({ message: 'Login Successful!' });
-    });
+  app.post('/auth/login', login, jwtSign, (req, res) => {
+    res.status(200).json({ message: 'Login Successful!' });
   });
 };
